backend/requirements.txt
fastapi
uvicorn[standard]
black
llama-index-llms-openai
llama-index